+++ CHANGELOG.md
@@ -321,3 +321,4 @@
 - [2026-04-11 16:23] refactor(security): optimize data flow between domain and data layers
 - [2026-04-11 20:54] refactor(analytics): optimize data flow between domain and data layers
 - [2026-04-11 21:15] feat(analytics): implement shimmer loading skeleton in analytics
+- [2026-04-11 21:25] feat(storage): add bi-directional event bus handler for storage